fix unsigned int to int comparisons

+++ Cards/main.cpp
@@ -25,7 +25,7 @@ int main()
 	p4.sortHand();
 	std::string suitName = "";
 	std::cout << p1.getName() << "'s hand: " << std::endl;
-	for (int i = 0; i < p1.hand.size(); i++)
+	for (int i = 0; i < (int)p1.hand.size(); i++)
 	{
 		if (p1.hand[i].getSuit() == 0) suitName = "CLUBS";
 		else if (p1.hand[i].getSuit() == 1) suitName = "HEARTS";
@@ -34,7 +34,7 @@ int main()
 		std::cout << p1.hand[i].getId() << ' ' << p1.hand[i].getValue() << ' ' << suitName << std::endl;
 	}
 	std::cout << p2.getName() << "'s hand: " << std::endl;
-	for (int i = 0; i < p2.hand.size(); i++)
+	for (int i = 0; i < (int)p2.hand.size(); i++)
 	{
 		if (p2.hand[i].getSuit() == 0) suitName = "CLUBS";
 		else if (p2.hand[i].getSuit() == 1) suitName = "HEARTS";
@@ -43,7 +43,7 @@ int main()
 		std::cout << p2.hand[i].getId() << ' ' << p2.hand[i].getValue() << ' ' << suitName << std::endl;
 	}
 	std::cout << p3.getName() << "'s hand: " << std::endl;
-	for (int i = 0; i < p1.hand.size(); i++)
+	for (int i = 0; i < (int)p1.hand.size(); i++)
 	{
 		if (p3.hand[i].getSuit() == 0) suitName = "CLUBS";
 		else if (p3.hand[i].getSuit() == 1) suitName = "HEARTS";
@@ -52,7 +52,7 @@ int main()
 		std::cout << p3.hand[i].getId() << ' ' << p3.hand[i].getValue() << ' ' << suitName << std::endl;
 	}
 	std::cout << p4.getName() << "'s hand: " << std::endl;
-	for (int i = 0; i < p4.hand.size(); i++)
+	for (int i = 0; i < (int)p4.hand.size(); i++)
 	{
 		if (p4.hand[i].getSuit() == 0) suitName = "CLUBS";
 		else if (p4.hand[i].getSuit() == 1) suitName = "HEARTS";
